requirements.txt
pg
iexfinance
python-dotenv